+++ Dtos/User/UserDetailsDTO.cs
@@ -12,7 +12,7 @@ namespace StockMarket.Dtos.User
         public string Email { get; set; } = string.Empty;
         public DateTime CreatedAt { get; set; } = DateTime.Now;
         public int RoleId { get; set; }
-        public int? Balance { get; set; }
+        public decimal? Balance { get; set; }
         public int? PortfolioId { get; set; }
         public int? TotalStocks { get; set; }
         public int? TotalStockQuantity { get; set; }
